dfxp: data structures fixup
I arranged the data structures in a more-meaningful way. For helper
functions, I used the K&R brace style for two reasons:

(1) It helps the reader identify which code is ours
(2) It takes up less space on the page

We will have to convert back to Alman style, but this shouldn't be
hard. There was an error in the display alignment table because it
was just a copy+paste of the textalign table. I fixed that.

I also used enums to help the reader understand what the tables
meant. These tables values are from the captions sample PDF
ending in 610.pdf