dfxp: create stack for style information (commented out)
Created the style stack along with the TODO(s) for what to do with
it. We want to keep track of the current style associated with the
parent tags by making a stack alongside the nodestack. The style
stack stores a pointer to the node it was found in as well as the
style information itself (as a value).

When we ascend, we check whether the element we ascended to
matches the style in the style stack. When this happens, pop it
off the stack.

When we descend, we check whether the new element has style
information and use it. This is more complicated, first we need
to check if there is a region defined and use that as the base style
info. Next, we blend that together with any local style elements or
decorations.

The stack is currently commented out.