dfxp: layout & styling: helper to parse the layout and style metadata
The layout and styling elements form the basis of alignment and
decoration settings defined by a dfxp document. This is the start
of a helper function that will extract this metadata from the elements.

The styling block declares styles, while the layout block declares
regions. Regions can refer to styles by id or contain embedded style
blocks. If a region has an embedded style block, it overwrites those
explicitly-defined properties.