dfxp: stash style/layout helper function elsewhere
Upon further examination of the Java implementation, we discovered
that the Viacom captions parser does not consider origin/extent or
parse new layout. While good to have, it would not be a good use of
time to implement this part right now, so I'm deleting the skeleton
code I prepared for it in order to commit a clean/compilable state.

Instead of parsing the layout/styles, we will detect Viacom styles
and hard-code them for now. It is unlikely that other captions we have
use and depend on "defaultSpeaker" regions.

In the future, we can complete this function after deciding how to
reconcile the behavior Viacom parser depends on.