dfxp: remove redundant length computing function
We replace calls to dfxp_append_string with a function variable of
our choice and create a fake append function that simply computes
the vod_strlen sum of all of its inputs.

	static u_char*
	dfxp_append_string(u_char* p, u_char* s)
	{
		while (*s) *p++ = *s++;
		return p;
	}

	static u_char*
	dfxp_fake_append_string(u_char* p, u_char* s)
	{
		return p + vod_strlen(s);
	}

The first call used to compute the size uses the fake append function:
	alloc_size = dfxp_append_text_content(node, 0, style->flag.decoration, dfxp_fake_append_string);

The subsequent one uses the real append function
	end = dfxp_append_text_content(node, end, style->flag.decoration, dfxp_append_string);

The copy/paste sizing function is then deleted.