dfxp: proof-of-concept that applies text decorations, text alignment, and display alignment (#3)
* dfxp: look for timestamp in div

* dfxp: proof-of-concept that applies text decorations in <span> tags

* dfxp: tablize

* dfxp: document functions

* dfxp: add more decoration overhead (this will need to be dynamic)

* dfxp: wip: decoration, style, regions

* dfxp: demonstrate that its possible to append after the cue point

* dfxp: dfxp_get_frame_body: document requirements passed for dfxp_append_text_content

* dfxp: table out dfxp->webvtt idioms

These rules are outlined in the "PDF"

Conversion details
Block <metadata/> after conversion will be lost.
Supported TTML style cues are:
TTML style cue Converts to
fontStyle="italic" <i/>
fontWeight="bold" <b/>
textDecoration="underline" <u/>
textAlign="start" align=start position=15%
textAlign="left" align=start position=15%
textAlign="center" align=middle position=50%
textAlign="right" align=end position=85%
textAlign="end" align=end position=85%
displayAlign="before" line=10%
displayAlign="center" line=50%
displayAlign="after" line=100%
There is TTML style, which description hard coded into the CCS

* dfxp: data structures fixup

I arranged the data structures in a more-meaningful way. For helper
functions, I used the K&R brace style for two reasons:

(1) It helps the reader identify which code is ours
(2) It takes up less space on the page

We will have to convert back to Alman style, but this shouldn't be
hard. There was an error in the display alignment table because it
was just a copy+paste of the textalign table. I fixed that.

I also used enums to help the reader understand what the tables
meant. These tables values are from the captions sample PDF
ending in 610.pdf

* dfxp: merge function: only merge non-null alignments

Also, added dfxp_can_contain_style and modified function signatures

* dfxp: create stack for style information (commented out)

Created the style stack along with the TODO(s) for what to do with
it. We want to keep track of the current style associated with the
parent tags by making a stack alongside the nodestack. The style
stack stores a pointer to the node it was found in as well as the
style information itself (as a value).

When we ascend, we check whether the element we ascended to
matches the style in the style stack. When this happens, pop it
off the stack.

When we descend, we check whether the new element has style
information and use it. This is more complicated, first we need
to check if there is a region defined and use that as the base style
info. Next, we blend that together with any local style elements or
decorations.

The stack is currently commented out.

* dfxp: add dfxp_append_style, fix warnings, deco flags

This example should compile and run; it doesn't use the style stack
and there is a style declared in the parse frames function for testing
purposes.

* dfxp: fix trimming behavior and dont append <i><u><b> to whitepsace

* dfxp: layout & styling: helper to parse the layout and style metadata

The layout and styling elements form the basis of alignment and
decoration settings defined by a dfxp document. This is the start
of a helper function that will extract this metadata from the elements.

The styling block declares styles, while the layout block declares
regions. Regions can refer to styles by id or contain embedded style
blocks. If a region has an embedded style block, it overwrites those
explicitly-defined properties.

* dfxp: stash style/layout helper function elsewhere

Upon further examination of the Java implementation, we discovered
that the Viacom captions parser does not consider origin/extent or
parse new layout. While good to have, it would not be a good use of
time to implement this part right now, so I'm deleting the skeleton
code I prepared for it in order to commit a clean/compilable state.

Instead of parsing the layout/styles, we will detect Viacom styles
and hard-code them for now. It is unlikely that other captions we have
use and depend on "defaultSpeaker" regions.

In the future, we can complete this function after deciding how to
reconcile the behavior Viacom parser depends on.

* dfxp: webvtt uses colons, not equal signs

Also, attach size parameter for textAlign right and textAlign end.

* dfxp: make flags easy to compare for simpler stack traversal ops

Since we are using the hard-coded regions and styles, it would be
better to keep the style data inside flags, minimizing the number of
strings and pointer copies we have to make when mixing variant
styles.

I also renamed some variables to make things easier to understand.

* dfxp: style stack: span

This is a stack that tracks nested spans. Viacom uses nested
spans in one of their sample files, so we have to support it. The
standard is conflicted over whether nested spans should be allowed.

NOTE: need to remove old logic now that stack exists

* dfxp: nesting of spans

* dfxp: style: make node stack into node+style stack

Adds an additional 24 bytes of overhead for each stack element, with
the current max depth of 10, this is less than 1KiB per frame of temp
data.

* dfxp: make stack work for display and text alignments

Also adds some temporary debug statements for testing.

* dfxp: correctly coalesce redundant tags across parent->child nodes

```
	<p begin='00:00:01.535' end='00:00:02.035' tts:textAlign='left' space='preserve' tts:fontWeight='bold'>
	<span style='block' tts:fontStyle='italic'>Outer text<span tts:textDecoration='underline'>inner text</span> end of outer text</span>
	</p>
```

```
00:00:01.535 --> 00:00:02.035 align:start position:15%
<b><i>Outer text<u>inner text</u> end of outer text</b></i>
```

* remove debugging statements

* dfxp: remove redundant length computing function

We replace calls to dfxp_append_string with a function variable of
our choice and create a fake append function that simply computes
the vod_strlen sum of all of its inputs.

	static u_char*
	dfxp_append_string(u_char* p, u_char* s)
	{
		while (*s) *p++ = *s++;
		return p;
	}

	static u_char*
	dfxp_fake_append_string(u_char* p, u_char* s)
	{
		return p + vod_strlen(s);
	}

The first call used to compute the size uses the fake append function:
	alloc_size = dfxp_append_text_content(node, 0, style->flag.decoration, dfxp_fake_append_string);

The subsequent one uses the real append function
	end = dfxp_append_text_content(node, end, style->flag.decoration, dfxp_append_string);

The copy/paste sizing function is then deleted.

* dfxp: decorations: reverse close tag traversal order + cleanup

Before: <b><i>Text</b></i>
After: <b><i>Text</i></b>

* dfxp: fix style inheritance from child elements

Some elaborate testing reveals a bug in the current style inheritance logic. We should be setting the current node's style to the parent's style and then excising any style information from that current node.

* dfxp: viacom: formatting: rearrange size,position,align order

* dfxp: cleanup suggestions

* dfxp: remove leading spaces from captions